+++ src/lilypad/types/ee/projects/annotation_create_params.py
@@ -5,9 +5,6 @@ from __future__ import annotations
 from typing import List, Iterable, Optional
 from typing_extensions import Required, TypedDict
 
-from .label import Label
-from .evaluation_type import EvaluationType
-
 __all__ = ["AnnotationCreateParams", "Body"]
 
 
@@ -18,18 +15,10 @@ class AnnotationCreateParams(TypedDict, total=False):
 class Body(TypedDict, total=False):
     assigned_to: Optional[List[str]]
 
-    data: Optional[object]
+    assignee_email: Optional[str]
 
     function_uuid: Optional[str]
 
-    label: Optional[Label]
-    """Label enum"""
-
     project_uuid: Optional[str]
 
-    reasoning: Optional[str]
-
     span_uuid: Optional[str]
-
-    type: Optional[EvaluationType]
-    """Evaluation type enum"""