+++ src/lilypad/types/ee/projects/function_run_playground_response.py
@@ -1,7 +1,23 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-from typing_extensions import TypeAlias
+from typing import Optional
 
-__all__ = ["FunctionRunPlaygroundResponse"]
+from ...._models import BaseModel
 
-FunctionRunPlaygroundResponse: TypeAlias = str
+__all__ = ["FunctionRunPlaygroundResponse", "TraceContext"]
+
+
+class TraceContext(BaseModel):
+    span_uuid: Optional[str] = None
+    """The unique identifier for the current span within the trace."""
+
+
+class FunctionRunPlaygroundResponse(BaseModel):
+    result: object
+    """The result returned by the executed function.
+
+    Can be any JSON-serializable type.
+    """
+
+    trace_context: Optional[TraceContext] = None
+    """Represents the tracing context information provided by Lilypad."""