+++ tests/api_resources/ee/projects/test_functions.py
@@ -9,8 +9,8 @@ import pytest
 
 from lilypad import Lilypad, AsyncLilypad
 from tests.utils import assert_matches_type
-from lilypad._utils import parse_datetime
 from lilypad.types.ee.projects import (
+    FunctionRunPlaygroundResponse,
     FunctionGetAnnotationsResponse,
 )
 
@@ -78,11 +78,14 @@ class TestFunctions:
         function = client.ee.projects.functions.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
+            call_params={},
             model="model",
+            prompt_template="prompt_template",
             provider="openai",
         )
-        assert_matches_type(str, function, path=["response"])
+        assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
     @pytest.mark.skip()
     @parametrize
@@ -90,41 +93,22 @@ class TestFunctions:
         function = client.ee.projects.functions.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
-            model="model",
-            provider="openai",
-            function={
-                "code": "code",
-                "hash": "hash",
-                "name": "x",
-                "signature": "signature",
-                "archived": parse_datetime("2019-12-27T18:11:19.117Z"),
-                "arg_types": {"foo": "string"},
-                "call_params": {
-                    "frequency_penalty": 0,
-                    "max_tokens": 0,
-                    "presence_penalty": 0,
-                    "seed": 0,
-                    "stop": "string",
-                    "temperature": 0,
-                    "top_p": 0,
-                },
-                "custom_id": "custom_id",
-                "dependencies": {
-                    "foo": {
-                        "extras": ["string"],
-                        "version": "version",
-                    }
-                },
-                "is_versioned": True,
-                "model": "model",
-                "project_uuid": "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
-                "prompt_template": "prompt_template",
-                "provider": "provider",
-                "version_num": 0,
+            call_params={
+                "frequency_penalty": 0,
+                "max_tokens": 0,
+                "presence_penalty": 0,
+                "seed": 0,
+                "stop": "string",
+                "temperature": 0,
+                "top_p": 0,
             },
+            model="model",
+            prompt_template="prompt_template",
+            provider="openai",
         )
-        assert_matches_type(str, function, path=["response"])
+        assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
     @pytest.mark.skip()
     @parametrize
@@ -132,15 +116,18 @@ class TestFunctions:
         response = client.ee.projects.functions.with_raw_response.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
+            call_params={},
             model="model",
+            prompt_template="prompt_template",
             provider="openai",
         )
 
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         function = response.parse()
-        assert_matches_type(str, function, path=["response"])
+        assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
     @pytest.mark.skip()
     @parametrize
@@ -148,15 +135,18 @@ class TestFunctions:
         with client.ee.projects.functions.with_streaming_response.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
+            call_params={},
             model="model",
+            prompt_template="prompt_template",
             provider="openai",
         ) as response:
             assert not response.is_closed
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             function = response.parse()
-            assert_matches_type(str, function, path=["response"])
+            assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -167,8 +157,11 @@ class TestFunctions:
             client.ee.projects.functions.with_raw_response.run_playground(
                 function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
                 project_uuid="",
+                arg_types={"foo": "string"},
                 arg_values={"foo": 0},
+                call_params={},
                 model="model",
+                prompt_template="prompt_template",
                 provider="openai",
             )
 
@@ -176,8 +169,11 @@ class TestFunctions:
             client.ee.projects.functions.with_raw_response.run_playground(
                 function_uuid="",
                 project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+                arg_types={"foo": "string"},
                 arg_values={"foo": 0},
+                call_params={},
                 model="model",
+                prompt_template="prompt_template",
                 provider="openai",
             )
 
@@ -243,11 +239,14 @@ class TestAsyncFunctions:
         function = await async_client.ee.projects.functions.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
+            call_params={},
             model="model",
+            prompt_template="prompt_template",
             provider="openai",
         )
-        assert_matches_type(str, function, path=["response"])
+        assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
     @pytest.mark.skip()
     @parametrize
@@ -255,41 +254,22 @@ class TestAsyncFunctions:
         function = await async_client.ee.projects.functions.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
-            model="model",
-            provider="openai",
-            function={
-                "code": "code",
-                "hash": "hash",
-                "name": "x",
-                "signature": "signature",
-                "archived": parse_datetime("2019-12-27T18:11:19.117Z"),
-                "arg_types": {"foo": "string"},
-                "call_params": {
-                    "frequency_penalty": 0,
-                    "max_tokens": 0,
-                    "presence_penalty": 0,
-                    "seed": 0,
-                    "stop": "string",
-                    "temperature": 0,
-                    "top_p": 0,
-                },
-                "custom_id": "custom_id",
-                "dependencies": {
-                    "foo": {
-                        "extras": ["string"],
-                        "version": "version",
-                    }
-                },
-                "is_versioned": True,
-                "model": "model",
-                "project_uuid": "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
-                "prompt_template": "prompt_template",
-                "provider": "provider",
-                "version_num": 0,
+            call_params={
+                "frequency_penalty": 0,
+                "max_tokens": 0,
+                "presence_penalty": 0,
+                "seed": 0,
+                "stop": "string",
+                "temperature": 0,
+                "top_p": 0,
             },
+            model="model",
+            prompt_template="prompt_template",
+            provider="openai",
         )
-        assert_matches_type(str, function, path=["response"])
+        assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
     @pytest.mark.skip()
     @parametrize
@@ -297,15 +277,18 @@ class TestAsyncFunctions:
         response = await async_client.ee.projects.functions.with_raw_response.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
+            call_params={},
             model="model",
+            prompt_template="prompt_template",
             provider="openai",
         )
 
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         function = await response.parse()
-        assert_matches_type(str, function, path=["response"])
+        assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
     @pytest.mark.skip()
     @parametrize
@@ -313,15 +296,18 @@ class TestAsyncFunctions:
         async with async_client.ee.projects.functions.with_streaming_response.run_playground(
             function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+            arg_types={"foo": "string"},
             arg_values={"foo": 0},
+            call_params={},
             model="model",
+            prompt_template="prompt_template",
             provider="openai",
         ) as response:
             assert not response.is_closed
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             function = await response.parse()
-            assert_matches_type(str, function, path=["response"])
+            assert_matches_type(FunctionRunPlaygroundResponse, function, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -332,8 +318,11 @@ class TestAsyncFunctions:
             await async_client.ee.projects.functions.with_raw_response.run_playground(
                 function_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
                 project_uuid="",
+                arg_types={"foo": "string"},
                 arg_values={"foo": 0},
+                call_params={},
                 model="model",
+                prompt_template="prompt_template",
                 provider="openai",
             )
 
@@ -341,7 +330,10 @@ class TestAsyncFunctions:
             await async_client.ee.projects.functions.with_raw_response.run_playground(
                 function_uuid="",
                 project_uuid="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+                arg_types={"foo": "string"},
                 arg_values={"foo": 0},
+                call_params={},
                 model="model",
+                prompt_template="prompt_template",
                 provider="openai",
             )