+++ tests/api_resources/test_projects.py
@@ -13,6 +13,7 @@ from lilypad.types import (
     ProjectPublic,
     ProjectListResponse,
     ProjectDeleteResponse,
+    ProjectRetrieveTagsResponse,
 )
 from lilypad._utils import parse_datetime
 from lilypad.types.projects.functions import FunctionPublic
@@ -309,6 +310,48 @@ class TestProjects:
                 signature="signature",
             )
 
+    @pytest.mark.skip()
+    @parametrize
+    def test_method_retrieve_tags(self, client: Lilypad) -> None:
+        project = client.projects.retrieve_tags(
+            "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+        )
+        assert_matches_type(ProjectRetrieveTagsResponse, project, path=["response"])
+
+    @pytest.mark.skip()
+    @parametrize
+    def test_raw_response_retrieve_tags(self, client: Lilypad) -> None:
+        response = client.projects.with_raw_response.retrieve_tags(
+            "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+        )
+
+        assert response.is_closed is True
+        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+        project = response.parse()
+        assert_matches_type(ProjectRetrieveTagsResponse, project, path=["response"])
+
+    @pytest.mark.skip()
+    @parametrize
+    def test_streaming_response_retrieve_tags(self, client: Lilypad) -> None:
+        with client.projects.with_streaming_response.retrieve_tags(
+            "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+        ) as response:
+            assert not response.is_closed
+            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+
+            project = response.parse()
+            assert_matches_type(ProjectRetrieveTagsResponse, project, path=["response"])
+
+        assert cast(Any, response.is_closed) is True
+
+    @pytest.mark.skip()
+    @parametrize
+    def test_path_params_retrieve_tags(self, client: Lilypad) -> None:
+        with pytest.raises(ValueError, match=r"Expected a non-empty value for `project_uuid` but received ''"):
+            client.projects.with_raw_response.retrieve_tags(
+                "",
+            )
+
 
 class TestAsyncProjects:
     parametrize = pytest.mark.parametrize("async_client", [False, True], indirect=True, ids=["loose", "strict"])
@@ -597,4 +640,46 @@ class TestAsyncProjects:
                 hash="hash",
                 name="x",
                 signature="signature",
+            )
+
+    @pytest.mark.skip()
+    @parametrize
+    async def test_method_retrieve_tags(self, async_client: AsyncLilypad) -> None:
+        project = await async_client.projects.retrieve_tags(
+            "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+        )
+        assert_matches_type(ProjectRetrieveTagsResponse, project, path=["response"])
+
+    @pytest.mark.skip()
+    @parametrize
+    async def test_raw_response_retrieve_tags(self, async_client: AsyncLilypad) -> None:
+        response = await async_client.projects.with_raw_response.retrieve_tags(
+            "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+        )
+
+        assert response.is_closed is True
+        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+        project = await response.parse()
+        assert_matches_type(ProjectRetrieveTagsResponse, project, path=["response"])
+
+    @pytest.mark.skip()
+    @parametrize
+    async def test_streaming_response_retrieve_tags(self, async_client: AsyncLilypad) -> None:
+        async with async_client.projects.with_streaming_response.retrieve_tags(
+            "182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+        ) as response:
+            assert not response.is_closed
+            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+
+            project = await response.parse()
+            assert_matches_type(ProjectRetrieveTagsResponse, project, path=["response"])
+
+        assert cast(Any, response.is_closed) is True
+
+    @pytest.mark.skip()
+    @parametrize
+    async def test_path_params_retrieve_tags(self, async_client: AsyncLilypad) -> None:
+        with pytest.raises(ValueError, match=r"Expected a non-empty value for `project_uuid` but received ''"):
+            await async_client.projects.with_raw_response.retrieve_tags(
+                "",
             )