+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 0.1.0-alpha.20 (2025-04-15)
+
+Full Changelog: [v0.1.0-alpha.19...v0.1.0-alpha.20](https://github.com/Mirascope/lilypad-sdk-python/compare/v0.1.0-alpha.19...v0.1.0-alpha.20)
+
+### Features
+
+* **api:** update via SDK Studio ([c3a2fc4](https://github.com/Mirascope/lilypad-sdk-python/commit/c3a2fc47d1c0760994bade43ab2a75b548aecfb9))
+* **api:** update via SDK Studio ([168bfe5](https://github.com/Mirascope/lilypad-sdk-python/commit/168bfe5cd8ae1b5a829fdbfc12e967d9ffdfca3d))
+
+
+### Bug Fixes
+
+* **perf:** optimize some hot paths ([59b67ca](https://github.com/Mirascope/lilypad-sdk-python/commit/59b67ca05fac0077c1524874e9767a4faf47403f))
+* **perf:** skip traversing types for NotGiven values ([7367bbd](https://github.com/Mirascope/lilypad-sdk-python/commit/7367bbde711a0ebe76de10b17b40a6d2821b3851))
+
+
+### Chores
+
+* **client:** minor internal fixes ([14acfa7](https://github.com/Mirascope/lilypad-sdk-python/commit/14acfa70a0d175d31c3fd58ce2aa6196ea0f73ba))
+* **internal:** update pyright settings ([1ba41f6](https://github.com/Mirascope/lilypad-sdk-python/commit/1ba41f6ae04bacea78b589b39c534608720b0174))
+
+
 ## 0.1.0-alpha.19 (2025-04-11)
 
 Full Changelog: [v0.1.0-alpha.18...v0.1.0-alpha.19](https://github.com/Mirascope/lilypad-sdk-python/compare/v0.1.0-alpha.18...v0.1.0-alpha.19)